# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "مستندات API برای ثبت‌نام و احراز هویت کاربران",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/wallet/deposit": {
            "post": {
                "tags": [
                    "Wallet (User)"
                ],
                "summary": "شارژ کیف پول (User)",
                "operationId": "b3d556d44c664e880a8d4a9c7573797a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "maximum": 10000000,
                                        "minimum": 1000,
                                        "example": 50000
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Wallet deposit"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deposit successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "شارژ انجام شد"
                                        },
                                        "balance": {
                                            "type": "integer",
                                            "example": 50000
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid amount",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "مبلغ واریزی باید بین 1000 تا 10,000,000 تومان باشد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/wallet/use-gift-card": {
            "post": {
                "tags": [
                    "Wallet (User)"
                ],
                "summary": "Use a gift card to add balance to wallet (User)",
                "description": "Use gift card",
                "operationId": "1c3f29e0a0a2605801755d475860940a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "GIFT123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gift card used successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "کارت هدیه با موفقیت استفاده شد"
                                        },
                                        "balance": {
                                            "type": "integer",
                                            "example": 50000
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid gift card",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "کارت هدیه نامعتبر است"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/wallet/transactions": {
            "get": {
                "tags": [
                    "Wallet (User)"
                ],
                "summary": "لیست تراکنش‌های کیف پول (User)",
                "operationId": "1e0bf844dcafd94420ba2d60fe4d73ac",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/WalletTransactionResource"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/wallet/withdraw": {
            "post": {
                "tags": [
                    "Wallet (User)"
                ],
                "summary": "درخواست برداشت از کیف پول (User)",
                "operationId": "e0d3936a45afa61dcafe3ac04505a130",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "bank_card_id"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "maximum": 10000000,
                                        "minimum": 1000,
                                        "example": 50000
                                    },
                                    "bank_card_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Withdrawal request submitted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "درخواست برداشت ثبت شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "موجودی کافی نیست"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "WalletTransactionResource": {
                "title": "Wallet Transaction Resource",
                "description": "منبع تراکنش کیف پول شامل اطلاعات تراکنش",
                "properties": {
                    "id": {
                        "description": "شناسه تراکنش",
                        "type": "integer",
                        "example": 1
                    },
                    "amount": {
                        "description": "مبلغ تراکنش به تومان",
                        "type": "integer",
                        "example": 50000
                    },
                    "type": {
                        "description": "نوع تراکنش",
                        "type": "string",
                        "enum": [
                            "deposit",
                            "withdraw",
                            "transfer",
                            "deposit_gift_card"
                        ],
                        "example": "deposit"
                    },
                    "description": {
                        "description": "توضیحات تراکنش",
                        "type": "string",
                        "example": "شارژ کیف پول",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "تاریخ ایجاد تراکنش",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-23 10:00:00"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}